fix(dashboard): ship protection-stats chunk for audit workspace
Include the generated supply-chain-protection-stats static chunk so audit
workspace can resolve manager coverage at runtime.
