+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/policy-workspace.js
@@ -1,4 +1,4 @@
-import { r as reactExports, b6 as savePolicyDecision, j as jsxRuntimeExports, S as SectionLabel, h as harnessDisplayName, A as ActionButton, b7 as HiMiniChevronLeft, y as HiMiniChevronRight, b8 as policyActionLabel, b9 as scopeLabel, b as EmptyState, ac as Tag, p as HiMiniChevronUp, q as HiMiniChevronDown, B as Badge, m as formatRelativeTime, b4 as guardAwareHref, a$ as HiMiniCloudArrowUp, ax as HiMiniTrash, ad as HiMiniMagnifyingGlass, ba as HiMiniPlus } from "../guard-dashboard.js";
+import { r as reactExports, b7 as savePolicyDecision, j as jsxRuntimeExports, S as SectionLabel, h as harnessDisplayName, A as ActionButton, b8 as HiMiniChevronLeft, y as HiMiniChevronRight, b9 as policyActionLabel, ba as scopeLabel, b as EmptyState, ac as Tag, p as HiMiniChevronUp, q as HiMiniChevronDown, B as Badge, m as formatRelativeTime, b5 as guardAwareHref, b0 as HiMiniCloudArrowUp, ax as HiMiniTrash, ad as HiMiniMagnifyingGlass, bb as HiMiniPlus } from "../guard-dashboard.js";
 const ACTION_FAMILIES = [
   { id: "package-request", label: "Package installs", example: "npm, pip, pnpm installs" },
   { id: "tool-action", label: "Shell and tool commands", example: "terminal commands agents run" },
